renamed cdftools flags, update rhino platform settings incl. default PRIMAVERA filters

+++ hiresclim2/script/nemo_post.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-set -ex
+set -e
 
  ############################################
  # To be called from ../master_hiresclim.sh #
@@ -38,10 +38,12 @@ NOMP=${NEMO_NPROCS}
 echo "SBC file used: ${use_SBC:=1}"
 (( $use_SBC )) && SBC='SBC' || SBC='grid_T'
 
-# Check if new or old version of CDFtools is used - can be set in
-# your ../../../conf/conf_hiresclim_<MACHINE-NAME>.sh, but old syntax
-# is assumed if not set.
-echo "Use newer syntax for cdfmean: ${newercdftools:=0}"
+# Check which version of CDFtools is used - can be set in your
+# ../../../conf/conf_hiresclim_<MACHINE-NAME>.sh, but old syntax is
+# assumed if not set.
+echo "Use CDFTOOLS 4.x syntax: ${cdftools4:=0}"
+echo "Use CDFTOOLS 3.0.1 syntax: ${cdftools301:=0}"
+if (( $cdftools4 )); then cdftools301=1; fi
 
 # update NEMORESULTS and get OUTDIR0
 eval_dirs 1
@@ -158,7 +160,7 @@ if [ ${ivf} -eq 1 ]; then
 fi
 
 # ICE
-if (( $newercdftools ))         # auxilliary file for newer CDFtools (4.0 master retrieved on 06-09-2017)
+if (( $cdftools4 ))         # auxilliary file for newer CDFtools (4.0 master retrieved on 06-09-2017)
 then
     if [ "${nm_iceconc}" != "siconc" ]; then
         ncrename -v ${nm_iceconc},siconc  ${froot}_icemod.nc  ${froot}_icemod_cdfnew.nc
@@ -172,17 +174,19 @@ fi
 if [ "${nm_iceconc}" != "iiceconc" ]; then ncrename -v ${nm_iceconc},iiceconc  ${froot}_icemod.nc ; fi
 if [ "${nm_icethic}" != "iicethic" ]; then ncrename -v ${nm_icethic},iicethic  ${froot}_icemod.nc ; fi
 
-# SHACONEMO update (april 2018) changes dimension names in the icemod file
-if (( $newercdftools ))
+# SHACONEMO update (april 2018) changes dimension names in the icemod files
+if ! ncdump -h ${froot}_icemod.nc | grep -q "^[[:blank:]]*x *="
 then
-    ncks -3 ${froot}_icemod_cdfnew.nc ${froot}_icemod_tmp.nc
-    ncrename -O -d .x_grid_T,x -d .y_grid_T,y ${froot}_icemod_tmp.nc ${froot}_icemod_cdfnew.nc
-    rm -f ${froot}_icemod_tmp1.nc
+    if (( $cdftools4 ))
+    then
+        ncks -3 ${froot}_icemod_cdfnew.nc ${froot}_icemod_tmp.nc
+        ncrename -O -d .x_grid_T,x -d .y_grid_T,y ${froot}_icemod_tmp.nc ${froot}_icemod_cdfnew.nc
+        rm -f ${froot}_icemod_tmp.nc
+    fi
+    ncks -3 ${froot}_icemod.nc ${froot}_icemod_tmp.nc
+    ncrename -O -d .x_grid_T,x -d .y_grid_T,y ${froot}_icemod_tmp.nc ${froot}_icemod.nc
+    rm -f ${froot}_icemod_tmp.nc
 fi
-ncks -3 ${froot}_icemod.nc ${froot}_icemod_tmp.nc
-ncrename -O -d .x_grid_T,x -d .y_grid_T,y ${froot}_icemod_tmp.nc ${froot}_icemod.nc
-rm -f ${froot}_icemod_tmp.nc
-
 
 # create time axis
 $cdo showdate ${froot}_icemod.nc | tr '[:blank:]' '\n' | \
@@ -202,7 +206,7 @@ rm -f $tempf
 for v in sosstsst sosaline sossheig # -- rename if needed, and average
 do
    [[ -f ${out}_${v}.nc4 ]] && mv ${out}_${v}.nc4 ${out}_${v}.nc
-   (( $newercdftools )) && $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v $v -p T \
+   (( $cdftools4 )) && $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v $v -p T \
            || $cdftoolsbin/cdfmean ${froot}_grid_T.nc $v T
    #$cdozip -selvar,mean_$v cdfmean.nc  ${out}_${v}_mean.nc
    $cdozip copy cdfmean.nc  ${out}_${v}_mean.nc
@@ -210,7 +214,7 @@ done
 for v in sowaflup
 do
    [[ -f ${out}_${v}.nc4 ]] && mv ${out}_${v}.nc4 ${out}_${v}.nc
-   (( $newercdftools )) && $cdftoolsbin/cdfmean -f ${froot}_${SBC}.nc -v $v -p T \
+   (( $cdftools4 )) && $cdftoolsbin/cdfmean -f ${froot}_${SBC}.nc -v $v -p T \
            || $cdftoolsbin/cdfmean ${froot}_${SBC}.nc $v T
    #$cdozip -selvar,mean_$v cdfmean.nc  ${out}_${v}_mean.nc
    $cdozip copy cdfmean.nc  ${out}_${v}_mean.nc
@@ -219,7 +223,7 @@ done
 
 # save global salinity and temperature mean
 for v in votemper vosaline; do
-    (( $newercdftools )) && $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v $v -p T \
+    (( $cdftools4 )) && $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v $v -p T \
             || $cdftoolsbin/cdfmean ${froot}_grid_T.nc $v T 
     #$cdozip -selvar,mean_$v,mean_3D$v cdfmean.nc  ${out}_${v}_mean.nc
     $cdozip copy cdfmean.nc  ${out}_${v}_mean.nc
@@ -238,7 +242,7 @@ for v in iiceconc iicethic; do
     if [ -f ${ff}4 ]; then mv ${ff}4 ${ff}; fi
 done
 
-if (( $newercdftools ))
+if (( $cdftools4 ))
 then
     $cdozip selvar,iiceconc,iicethic ${froot}_icemod.nc ${out}_ice.nc    
     $cdftoolsbin/cdficediags -i ${froot}_icemod_cdfnew.nc -lim3 -o ${out}_icediags.nc
@@ -249,11 +253,11 @@ else
 fi
 
     # ** MOC
-if (( $newercdftools ))
+if (( $cdftools4 ))
 then
-    #echo "FIXME! MOC with newercdftools"
-    
-    $cdftoolsbin/cdfmoc -v ${froot}_grid_V.nc -o ${out}_moc.nc
+    echo "FIXME! MOC with cdftools4"
+#forrtl: severe (174): SIGSEGV, segmentation fault occurred    
+    #$cdftoolsbin/cdfmoc -v ${froot}_grid_V.nc -o ${out}_moc.nc
 else
     $cdftoolsbin/cdfmoc ${froot}_grid_V.nc
     $cdozip copy moc.nc ${out}_moc.nc
@@ -261,125 +265,116 @@ fi
 
 if [[ $nemo_extra == 1 ]] ; then
 
-# ** heat content
-tmpstring=tmpdate
-for l in 0_300 300_800 800_bottom
-do
-    # when running cdftools 3.0.1, this is the only part where a newer syntax is required
-    if (( $newercdftools2 ))
-    then
-        eval "${cdftoolsbin}/cdfheatc -f ${froot}_grid_T.nc -zoom 0 0 0 0 \$depth_$l" | \
-            awk '/Total Heat content        :/ {print $5}' > tmp_$l
-    else
-        eval "${cdftoolsbin}/cdfheatc ${froot}_grid_T.nc 0 0 0 0 \$depth_$l" | \
-            awk '/Total Heat content        :/ {print $5}' > tmp_$l
-    fi
-    
-    tmpstring+=" tmp_$l"
-    $cdo -f nc settaxis,${year}-01-01,12:00:00,1mon -input,r1x1 ${out}_${l}_heatc.nc < tmp_$l
-done
-
-if (( $newercdftools ))
-then
-    # ** Nino3.4 SST
-    (( ! $skip_nino )) &&
-        $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v sosstsst -p T -w $nino34_region 0 0 -o ${out}_sosstsst_nino34.nc
-
-    # ** MOC
-#rhino: seg fault!    $cdftoolsbin/cdfmoc -v ${froot}_grid_V.nc -o ${out}_moc.nc
-
-    # barotropic stream function
-#rhino: seg fault!    $cdftoolsbin/cdfpsi -u ${froot}_grid_U.nc -v ${froot}_grid_V.nc -o ${out}_psi.nc
-
-    # mixed layer depth
-#rhino: seg fault!    $cdftoolsbin/cdfmxl -t ${froot}_grid_T.nc -o ${out}_mxl.nc
-
-else
-    # ** Nino3.4 SST
-    if (( ! $skip_nino )) ; then
-        $cdftoolsbin/cdfmean ${froot}_grid_T.nc sosstsst T $nino34_region 0 0
-        $cdozip copy cdfmean.nc  ${out}_sosstsst_nino34.nc
-    fi
-
-    # ** MOC
-    #$cdftoolsbin/cdfmoc ${froot}_grid_V.nc
-    #$cdozip copy moc.nc ${out}_moc.nc
-
-    # barotropic stream function
-    $cdftoolsbin/cdfpsi ${froot}_grid_U.nc ${froot}_grid_V.nc
-    $cdozip copy psi.nc ${out}_psi.nc
-
-    # mixed layer depth
-# TODO fix this (crashes with cdftools 3.0.1)
-#  File bathy_level.nc is missing 
-# Read mbathy in mesh_zgr.nc ...
-#forrtl: error (73): floating divide by zero
-    #$cdftoolsbin/cdfmxl ${froot}_grid_T.nc
-    #$cdozip copy mxl.nc ${out}_mxl.nc
-
-fi
-
-
-
-# TODO : add case for newer cdftools syntax    
-if [[ $newercdftools == 0 ]] ; then
-
-    #compute potential and in situ density
-    $cdftoolsbin/cdfsiginsitu ${froot}_grid_T.nc
-    $cdftoolsbin/cdfsig0 ${froot}_grid_T.nc
-    ncatted -a valid_min,vosigma0,m,f,0 -a valid_max,vosigma0,m,f,100 sig0.nc -o new_sig0.nc
-    ncatted -a valid_min,vosigmainsitu,m,f,0 -a valid_max,vosigmainsitu,m,f,100 siginsitu.nc -o new_siginsitu.nc
-
-    #0-350m upper ocean salinity and temperature
-    $cdftoolsbin/cdfvertmean ${froot}_grid_T.nc vosaline T 0 350
-    ncatted -a valid_max,vosaline_vert_mean,m,f,50 vertmean.nc -o new_vertmean.nc
-    $cdozip chname,vosaline_vert_mean,mlsaline -setctomiss,0 new_vertmean.nc ${out}_mlsaline.nc
-    rm new_vertmean.nc
-    $cdftoolsbin/cdfvertmean ${froot}_grid_T.nc votemper T 0 350
-    ncatted -a valid_min,votemper_vert_mean,m,f,-5 -a valid_max,votemper_vert_mean,m,f,50 vertmean.nc -o new_vertmean.nc
-    $cdozip chname,votemper_vert_mean,mltemper -setctomiss,0 new_vertmean.nc ${out}_mltemper.nc
-    rm new_vertmean.nc
-
-    #0-350m upper ocean density (potential and in-situ)
-    $cdftoolsbin/cdfvertmean new_siginsitu.nc vosigmainsitu T 0 350
-    $cdozip chname,vosigmainsitu_vert_mean,mlsigmas -setctomiss,0 vertmean.nc ${out}_mlsigmas.nc
-    $cdftoolsbin/cdfvertmean new_sig0.nc vosigma0 T 0 350
-    $cdozip chname,vosigma0_vert_mean,mlsigma0 -setctomiss,0 vertmean.nc ${out}_mlsigma0.nc
-
-    #surface currents
-    $cdftoolsbin/cdfvertmean ${froot}_grid_U.nc vozocrtx U 0 10
-    ncatted -a valid_min,vozocrtx_vert_mean,m,f,-10 -a valid_max,vozocrtx_vert_mean,m,f,10 vertmean.nc -o new_vertmean.nc
-    $cdozip chname,vozocrtx_vert_mean,sfczoncu -setctomiss,0 new_vertmean.nc ${out}_sfczoncu.nc
-    rm new_vertmean.nc
-    $cdftoolsbin/cdfvertmean ${froot}_grid_V.nc vomecrty V 0 10
-    ncatted -a valid_min,vomecrty_vert_mean,m,f,-10 -a valid_max,vomecrty_vert_mean,m,f,10 vertmean.nc -o new_vertmean.nc
-    $cdozip chname,vomecrty_vert_mean,sfcmercu -setctomiss,0 new_vertmean.nc ${out}_sfcmercu.nc
-
-    #zonally averaged profiles of temperatures and salinity
-    $cdftoolsbin/cdfzonalmean ${froot}_grid_T.nc T new_maskglo.nc
-    for vv in zotemper_glo zotemper_atl zotemper_inp zotemper_ind zotemper_pac \
-        zosaline_glo zosaline_atl zosaline_inp zosaline_ind zosaline_pac     ; do
-        ncatted -a valid_min,$vv,m,f,-10 -a valid_max,$vv,m,f,40 zonalmean.nc -o tmp_zonalmean.nc
-        mv tmp_zonalmean.nc zonalmean.nc ; rm -f tmp_zonalmean.nc
+    # ** heat content
+    tmpstring=tmpdate
+    for l in 0_300 300_800 800_bottom
+    do
+        # when running cdftools 3.0.1, this is the only part where a newer syntax is required
+        if (( $cdftools301 ))
+        then
+            eval "${cdftoolsbin}/cdfheatc -f ${froot}_grid_T.nc -zoom 0 0 0 0 \$depth_$l" | \
+                awk '/Total Heat content        :/ {print $5}' > tmp_$l
+        else
+            eval "${cdftoolsbin}/cdfheatc ${froot}_grid_T.nc 0 0 0 0 \$depth_$l" | \
+                awk '/Total Heat content        :/ {print $5}' > tmp_$l
+        fi
+        
+        tmpstring+=" tmp_$l"
+        $cdo -f nc settaxis,${year}-01-01,12:00:00,1mon -input,r1x1 ${out}_${l}_heatc.nc < tmp_$l
     done
-    $cdozip selvar,nav_lon,nav_lat,zotemper_glo,zotemper_atl,zotemper_inp,zotemper_ind,zotemper_pac zonalmean.nc ${out}_zotemper.nc
-    $cdozip selvar,nav_lon,nav_lat,zosaline_glo,zosaline_atl,zosaline_inp,zosaline_ind,zosaline_pac zonalmean.nc ${out}_zosaline.nc
 
-    #zonally averaged profiles of density
-    $cdftoolsbin/cdfzonalmean new_sig0.nc T new_maskglo.nc
-    $cdozip copy zonalmean.nc ${out}_zosigma0.nc
-    $cdftoolsbin/cdfzonalmean new_siginsitu.nc T new_maskglo.nc
-    $cdozip copy zonalmean.nc ${out}_zosigmas.nc
+    if (( $cdftools4 ))
+    then
+        # ** Nino3.4 SST
+        (( ! $skip_nino )) &&
+            $cdftoolsbin/cdfmean -f ${froot}_grid_T.nc -v sosstsst -p T -w $nino34_region 0 0 -o ${out}_sosstsst_nino34.nc
 
-    #regular-1grid-defined yearly-averaged multilevel T,S and sigma0 and sigmainsitu
-    $cdozip setctomiss,0 -yearmean -selvar,votemper,vosaline ${froot}_grid_T.nc temp0.nc
-    $cdozip remapbil,global_1 -setctomiss,0 -yearmean -setgrid,temp0.nc -selvar,vosigmainsitu new_siginsitu.nc temp1.nc
-    $cdozip -b 32 remapbil,global_1 temp0.nc temp2.nc
-    $cdozip remapbil,global_1 -setctomiss,0 -yearmean -setgrid,temp0.nc -selvar,vosigma0 new_sig0.nc temp3.nc
-    $cdozip merge temp1.nc temp2.nc temp3.nc fulloce.nc
-    mv fulloce.nc ${out}_fulloce.nc
+        # barotropic stream function
+        #rhino: seg fault!    $cdftoolsbin/cdfpsi -u ${froot}_grid_U.nc -v ${froot}_grid_V.nc -o ${out}_psi.nc
 
-fi # $newercdftools == 0
+        # mixed layer depth
+        #rhino: seg fault!    $cdftoolsbin/cdfmxl -t ${froot}_grid_T.nc -o ${out}_mxl.nc
+
+    else
+        # ** Nino3.4 SST
+        if (( ! $skip_nino )) ; then
+            $cdftoolsbin/cdfmean ${froot}_grid_T.nc sosstsst T $nino34_region 0 0
+            $cdozip copy cdfmean.nc  ${out}_sosstsst_nino34.nc
+        fi
+
+        # barotropic stream function
+        $cdftoolsbin/cdfpsi ${froot}_grid_U.nc ${froot}_grid_V.nc
+        $cdozip copy psi.nc ${out}_psi.nc
+
+        # mixed layer depth
+        # TODO fix this (crashes with cdftools 3.0.1)
+        #  File bathy_level.nc is missing 
+        # Read mbathy in mesh_zgr.nc ...
+        #forrtl: error (73): floating divide by zero
+        #$cdftoolsbin/cdfmxl ${froot}_grid_T.nc
+        #$cdozip copy mxl.nc ${out}_mxl.nc
+
+    fi
+
+    # TODO : add case for newer cdftools syntax    
+    if [[ $cdftools4 == 0 ]] ; then
+
+        #compute potential and in situ density
+        $cdftoolsbin/cdfsiginsitu ${froot}_grid_T.nc
+        $cdftoolsbin/cdfsig0 ${froot}_grid_T.nc
+        ncatted -a valid_min,vosigma0,m,f,0 -a valid_max,vosigma0,m,f,100 sig0.nc -o new_sig0.nc
+        ncatted -a valid_min,vosigmainsitu,m,f,0 -a valid_max,vosigmainsitu,m,f,100 siginsitu.nc -o new_siginsitu.nc
+
+        #0-350m upper ocean salinity and temperature
+        $cdftoolsbin/cdfvertmean ${froot}_grid_T.nc vosaline T 0 350
+        ncatted -a valid_max,vosaline_vert_mean,m,f,50 vertmean.nc -o new_vertmean.nc
+        $cdozip chname,vosaline_vert_mean,mlsaline -setctomiss,0 new_vertmean.nc ${out}_mlsaline.nc
+        rm new_vertmean.nc
+        $cdftoolsbin/cdfvertmean ${froot}_grid_T.nc votemper T 0 350
+        ncatted -a valid_min,votemper_vert_mean,m,f,-5 -a valid_max,votemper_vert_mean,m,f,50 vertmean.nc -o new_vertmean.nc
+        $cdozip chname,votemper_vert_mean,mltemper -setctomiss,0 new_vertmean.nc ${out}_mltemper.nc
+        rm new_vertmean.nc
+
+        #0-350m upper ocean density (potential and in-situ)
+        $cdftoolsbin/cdfvertmean new_siginsitu.nc vosigmainsitu T 0 350
+        $cdozip chname,vosigmainsitu_vert_mean,mlsigmas -setctomiss,0 vertmean.nc ${out}_mlsigmas.nc
+        $cdftoolsbin/cdfvertmean new_sig0.nc vosigma0 T 0 350
+        $cdozip chname,vosigma0_vert_mean,mlsigma0 -setctomiss,0 vertmean.nc ${out}_mlsigma0.nc
+
+        #surface currents
+        $cdftoolsbin/cdfvertmean ${froot}_grid_U.nc vozocrtx U 0 10
+        ncatted -a valid_min,vozocrtx_vert_mean,m,f,-10 -a valid_max,vozocrtx_vert_mean,m,f,10 vertmean.nc -o new_vertmean.nc
+        $cdozip chname,vozocrtx_vert_mean,sfczoncu -setctomiss,0 new_vertmean.nc ${out}_sfczoncu.nc
+        rm new_vertmean.nc
+        $cdftoolsbin/cdfvertmean ${froot}_grid_V.nc vomecrty V 0 10
+        ncatted -a valid_min,vomecrty_vert_mean,m,f,-10 -a valid_max,vomecrty_vert_mean,m,f,10 vertmean.nc -o new_vertmean.nc
+        $cdozip chname,vomecrty_vert_mean,sfcmercu -setctomiss,0 new_vertmean.nc ${out}_sfcmercu.nc
+
+        #zonally averaged profiles of temperatures and salinity
+        $cdftoolsbin/cdfzonalmean ${froot}_grid_T.nc T new_maskglo.nc
+        for vv in zotemper_glo zotemper_atl zotemper_inp zotemper_ind zotemper_pac \
+                               zosaline_glo zosaline_atl zosaline_inp zosaline_ind zosaline_pac     ; do
+            ncatted -a valid_min,$vv,m,f,-10 -a valid_max,$vv,m,f,40 zonalmean.nc -o tmp_zonalmean.nc
+            mv tmp_zonalmean.nc zonalmean.nc ; rm -f tmp_zonalmean.nc
+        done
+        $cdozip selvar,nav_lon,nav_lat,zotemper_glo,zotemper_atl,zotemper_inp,zotemper_ind,zotemper_pac zonalmean.nc ${out}_zotemper.nc
+        $cdozip selvar,nav_lon,nav_lat,zosaline_glo,zosaline_atl,zosaline_inp,zosaline_ind,zosaline_pac zonalmean.nc ${out}_zosaline.nc
+
+        #zonally averaged profiles of density
+        $cdftoolsbin/cdfzonalmean new_sig0.nc T new_maskglo.nc
+        $cdozip copy zonalmean.nc ${out}_zosigma0.nc
+        $cdftoolsbin/cdfzonalmean new_siginsitu.nc T new_maskglo.nc
+        $cdozip copy zonalmean.nc ${out}_zosigmas.nc
+
+        #regular-1grid-defined yearly-averaged multilevel T,S and sigma0 and sigmainsitu
+        $cdozip setctomiss,0 -yearmean -selvar,votemper,vosaline ${froot}_grid_T.nc temp0.nc
+        $cdozip remapbil,global_1 -setctomiss,0 -yearmean -setgrid,temp0.nc -selvar,vosigmainsitu new_siginsitu.nc temp1.nc
+        $cdozip -b 32 remapbil,global_1 temp0.nc temp2.nc
+        $cdozip remapbil,global_1 -setctomiss,0 -yearmean -setgrid,temp0.nc -selvar,vosigma0 new_sig0.nc temp3.nc
+        $cdozip merge temp1.nc temp2.nc temp3.nc fulloce.nc
+        mv fulloce.nc ${out}_fulloce.nc
+
+    fi # $cdftools4 == 0
 
 fi # $nemo_extra == 1
